pdf-csv/requirements.txt
tabula-py == 2.3.*
